de-duplicate wine module
